[SPARK-40667] move trait FileSourceOptionsSet to DataSourceOptions
