+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/FileSourceOptions.scala
@@ -40,44 +40,3 @@ object FileSourceOptions {
   val IGNORE_CORRUPT_FILES = "ignoreCorruptFiles"
   val IGNORE_MISSING_FILES = "ignoreMissingFiles"
 }
-
-/**
- * Interface defines for a file-based data source, how to
- *  - register a new option name
- *  - retrieve all registered option names
- *  - valid a given option name
- *  - get alternative option name if any
- */
-trait FileSourceOptionsSet {
-  // Option -> Alternative Option if any
-  private val validOptions = collection.mutable.Map[String, Option[String]]()
-
-  /**
-   * Register a new Option.
-   * @param name The primary option name
-   * @param alternative Alternative option name if any
-   */
-  protected def newOption(name: String, alternative: Option[String] = None): String = {
-    // Register both of the options
-    validOptions += (name -> alternative)
-    validOptions ++ alternative.map(alterName => alterName -> Some(name))
-    name
-  }
-
-  /**
-   * @return All valid file source options
-   */
-  def getAllValidOptionNames: scala.collection.Set[String] = validOptions.keySet
-
-  /**
-   * @param name Option name to be validated
-   * @return if the given Option name is valid
-   */
-  def isValidOptionName(name: String): Boolean = validOptions.contains(name)
-
-  /**
-   * @param name Option name
-   * @return Alternative option name if any
-   */
-  def getAlternativeOptionName(name: String): Option[String] = validOptions.getOrElse(name, None)
-}
