+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/json/JSONOptions.scala
@@ -25,7 +25,7 @@ import com.fasterxml.jackson.core.{JsonFactory, JsonFactoryBuilder}
 import com.fasterxml.jackson.core.json.JsonReadFeature
 
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.catalyst.{FileSourceOptions, FileSourceOptionsSet}
+import org.apache.spark.sql.catalyst.{DataSourceOptions, FileSourceOptions}
 import org.apache.spark.sql.catalyst.util._
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.internal.SQLConf.LegacyBehaviorPolicy
@@ -233,7 +233,7 @@ private[sql] object JSONOptionsInRead {
   )
 }
 
-object JSONOptions extends FileSourceOptionsSet {
+object JSONOptions extends DataSourceOptions {
   val SAMPLING_RATIO = newOption("samplingRatio")
   val PRIMITIVES_AS_STRING = newOption("primitivesAsString")
   val PREFERS_DECIMAL = newOption("prefersDecimal")
